requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
fastapi==0.110.0
filetype==1.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.65.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-core==1.2.17
langchain-google-genai==4.2.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.10
numpy==2.4.2
openai==2.24.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
pyparsing==3.3.2
python-dotenv==1.0.1
PyYAML==6.0.3
redis==7.2.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.36.3
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
